+++ crates/mimic/src/core/traits.rs
@@ -73,6 +73,7 @@ pub trait EntityKind: Kind + TypeKind + FieldValues {
     type PrimaryKey: Copy + Clone;
 
     const PRIMARY_KEY: &'static str;
+    const FIELDS: &'static [&'static str];
 
     fn key(&self) -> Key;
 }