+++ crates/mimic/src/db/query/delete.rs
@@ -1,6 +1,6 @@
 use crate::{
     core::{Value, traits::EntityKind},
-    db::query::{Cmp, FilterBuilder, FilterExpr},
+    db::query::{Cmp, FilterBuilder, FilterExpr, LimitExpr, QueryError, QueryValidate},
 };
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
@@ -12,6 +12,7 @@ use serde::{Deserialize, Serialize};
 #[derive(CandidType, Clone, Debug, Default, Deserialize, Serialize)]
 pub struct DeleteQuery {
     pub filter: Option<FilterExpr>,
+    pub limit: Option<LimitExpr>,
 }
 
 impl DeleteQuery {
@@ -40,7 +41,11 @@ impl DeleteQuery {
         self.filter_in(E::PRIMARY_KEY, Value::List(list))
     }
 
-    // filter
+    ///
+    /// FILTER
+    ///
+
+    /// Combines the existing filter expression with a new one using `And`.
     #[must_use]
     pub fn filter(mut self, f: impl FnOnce(FilterBuilder) -> FilterBuilder) -> Self {
         let builder = match self.filter.take() {
@@ -68,5 +73,36 @@ impl DeleteQuery {
     #[must_use]
     pub fn filter_in(self, field: &str, value: impl Into<Value>) -> Self {
         self.filter(|f| f.filter(field, Cmp::In, value))
+    }
+
+    ///
+    /// LIMIT
+    ///
+
+    #[must_use]
+    pub fn limit(mut self, limit: u32) -> Self {
+        let expr = self.limit.unwrap_or_default().limit(limit);
+        self.limit = Some(expr);
+        self
+    }
+
+    #[must_use]
+    pub fn offset(mut self, offset: u32) -> Self {
+        let expr = self.limit.unwrap_or_default().offset(offset);
+        self.limit = Some(expr);
+        self
+    }
+}
+
+impl<E: EntityKind> QueryValidate<E> for DeleteQuery {
+    fn validate(&self) -> Result<(), QueryError> {
+        if let Some(filter) = &self.filter {
+            QueryValidate::<E>::validate(filter)?;
+        }
+        if let Some(limit) = &self.limit {
+            QueryValidate::<E>::validate(limit)?;
+        }
+
+        Ok(())
     }
 }