+++ crates/mimic/src/db/query/limit.rs
@@ -1,0 +1,52 @@
+use crate::{
+    core::traits::EntityKind,
+    db::query::{QueryError, QueryValidate},
+};
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+
+///
+/// LimitExpr
+///
+
+#[derive(CandidType, Clone, Debug, Default, Deserialize, Serialize)]
+pub struct LimitExpr {
+    pub limit: Option<u32>,
+    pub offset: u32,
+}
+
+impl LimitExpr {
+    #[must_use]
+    pub fn new(limit: u32) -> Self {
+        Self {
+            limit: Some(limit),
+            offset: 0,
+        }
+    }
+
+    #[must_use]
+    pub fn with_offset(limit: u32, offset: u32) -> Self {
+        Self {
+            limit: Some(limit),
+            offset,
+        }
+    }
+
+    #[must_use]
+    pub fn limit(mut self, limit: u32) -> Self {
+        self.limit = Some(limit);
+        self
+    }
+
+    #[must_use]
+    pub fn offset(mut self, offset: u32) -> Self {
+        self.offset = offset;
+        self
+    }
+}
+
+impl<E: EntityKind> QueryValidate<E> for LimitExpr {
+    fn validate(&self) -> Result<(), QueryError> {
+        Ok(())
+    }
+}